+++ packages/core/src/themes/utils.ts
@@ -1,9 +1,7 @@
 import { cloneDeep, merge } from 'lodash'
-import { useTheme } from './context'
-import { ThemedComponent, ThemeSpec, WithVariant } from './types'
 import { CssProps } from '../general'
+import { ThemeSpec } from './types'
 import { camelCase } from './helpers'
-import { useMemo } from 'react'
 
 // construct a className string by composing a variant code and a className string
 export const getClassName = (
@@ -46,11 +44,4 @@ export const mergeThemes = (themes: ThemeSpec[]) => {
         result = merge(result, theme)
     })
     return result
-}
-
-export const useThemedProps = <Props extends WithVariant>(props: Props, key: ThemedComponent) => {
-    const theme = useTheme()
-    return useMemo(() => {
-        return merge(cloneDeep(theme[key][String(props.variant ?? 'default')]), props)
-    }, [theme, props, key])
 }