+++ packages/core/tests/themes/themes.utils.test.tsx
@@ -1,17 +1,13 @@
-import { render } from '@testing-library/react'
 import {
     getClassName,
     addColor,
     addOpacity,
     defaultTheme,
-    Chart,
-    AxisProps,
     CompleteThemeSpec,
     mergeTheme,
     ThemeSpec,
-    useThemedProps,
     mergeThemes,
-} from '../../src'
+} from '../../src/themes'
 
 describe('mergeTheme', () => {
     it('creates a copy', () => {
@@ -184,36 +180,5 @@ describe('addOpacity', () => {
     it('add to an empty style', () => {
         const result = addOpacity(undefined, 0.5)
         expect(result.opacity).toEqual(0.5)
-    })
-})
-
-describe('useThemedProps', () => {
-    it('completes props from theme', () => {
-        const customTheme: ThemeSpec = {
-            Axis: {
-                top: {
-                    offset: 10,
-                },
-                bottom: {
-                    offset: 5,
-                },
-                left: {
-                    offset: 7,
-                },
-            },
-        }
-        let result: AxisProps = { variant: 'top' }
-        const GetThemedAxisProps = (props: AxisProps) => {
-            result = useThemedProps(props, 'Axis')
-            return null
-        }
-
-        render(
-            <Chart theme={customTheme}>
-                <GetThemedAxisProps variant={'top'} />
-            </Chart>
-        )
-        expect(result['variant']).toEqual('top')
-        expect(result['offset']).toEqual(10)
     })
 })